pyproject.toml
[project]
name = "deployup"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.5.2",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]
